+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Ability to search.
 - Ability to use tags when creating a book.
+
+## [0.0.4] - 2022-08-06
+### Added
+- Ability to search.
+
+### Changed
+- DeletableType to ContentType.
 
 ## [0.0.3] - 2022-08-05
 ### Added